+++ app/community/page.tsx
@@ -9,7 +9,7 @@ import {
 } from "@/components/ui/empty";
 import { Header } from "@/components/layout/header";
 import { Instrument_Serif } from "next/font/google";
-import { WallpaperCard } from "@/components/wallpaper-card";
+import { WallpaperCard } from "@/components/misc/wallpaper-card";
 import prisma from "@/lib/prisma";
 
 const instrumentSerif = Instrument_Serif({
