+++ components/preview/wallpaper-canvas.tsx
